Fix cipher suite scanning and add cipher suite rating

+++ SSLTest/src/utils.py
@@ -28,7 +28,7 @@ def receive_data(sock, timeout, debug_source):
     Receive network data in chunks
 
     :param sock: Socket to receive from
-    :param int timeout: Timeout in seconds
+    :param float timeout: Timeout in seconds
     :param str debug_source: Description of the debug source
     :return: Array of bytes of received data
     :rtype: bytes
@@ -61,7 +61,7 @@ def send_data_return_sock(address, client_hello, timeout, debug_source):
 
     :param tuple address: Tuple of an url and port
     :param bytes client_hello: client_hello data in bytes
-    :param int timeout: Timeout in seconds
+    :param float timeout: Timeout in seconds
     :param str debug_source: Description of the debug source
     :return: Created socket and received response
     :rtype: bytes or socket
